fix collision error
fix inconsistency between how projectile positions are interpreted in collision detection vs drawing.
The collision detection was treating p.y as the center of the projectile, but the drawing function treats it as the top of the projectile.

Now treats p.y as the top of the projectile not the center